pyback/requirements.txt
aioredis==1.3.1
asgiref==3.2.10
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.6.2
Automat==20.2.0
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9.2
daphne==2.5.0
Django==3.0.8
django-appconf==1.0.4
django-cors-headers==3.4.0
django-filter==2.2.0
django-image-cropping==1.4.0
djangochannelsrestframework==0.0.5
djangorestframework==3.11.0
drf-yasg==1.17.1
easy-thumbnails==2.7
et-xmlfile==1.0.1
hiredis==1.0.1
html5lib==1.1
hyperlink==19.0.0
idna==2.10
importlib-metadata==1.7.0
incremental==17.5.0
inflection==0.5.0
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
MarkupSafe==1.1.1
more-itertools==8.4.0
msgpack==0.6.2
openpyxl==3.0.3
packaging==20.4
Pillow==7.1.1
pluggy==0.13.1
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pyparsing==2.4.7
pytest==5.4.0
pytest-asyncio==0.11.0
pytest-django==3.9.0
pytz==2020.1
requests==2.24.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
service-identity==18.1.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.4.1
uritemplate==3.0.1
urllib3==1.25.9
wcwidth==0.2.5
webencodings==0.5.1
zipp==3.1.0
zope.interface==5.1.0
